feat: added all the drops with usernames

+++ api/models/Purchase.js
@@ -22,7 +22,8 @@ Purchase.init(
     },
 
     pricePaid: {
-      type: DataTypes.INTEGER,
+      // type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(10, 2),
       allowNull: false,
     },
   },
